+++ src/audio/input/ffmpegBackend.h
@@ -145,7 +145,7 @@ public:
     bool seekable() const override { return true; }
 
     /// Seek specified position
-    bool seek(int pos) override;
+    bool seek(double pos) override;
 
     /// Get samplerate
     unsigned int samplerate() const override { return m_audioStream->codecpar->sample_rate; }
