+++ src/audio/input/sndBackend.h
@@ -93,7 +93,7 @@ public:
     bool seekable() const override { return true; }
 
     /// Seek specified position
-    bool seek(int pos) override;
+    bool seek(double pos) override;
 
     /// Get samplerate
     unsigned int samplerate() const override { return m_sf ? m_si.samplerate : 0; }
